Big changes
- Changed channel / model relationship, channels no longer define
  messages (Models) they support, but Messages subscribe to channels
- Implemented proper Model support with automatic Schema generation
- Actor: changed 'input_channels' and 'output_channels' to 'consume'
  and 'produces' which expects a tuple of Models which define the
  messages to 'subscribe' to

The changed channel/model relationship allows to automatically
deduce what messages are requested by the actor and which ones
are produced. This allows the actor to consume from the same channel
it produces to

It won't change anything about the topic idea, just the way how it is
expressed in the code.